option_atm_strategy/requirements.txt
kiteconnect
pandas
flask
python-dotenv
datetime
nsepy
yfinance
matplotlib
streamlit
streamlit_autorefresh